Add shell script to launch mac development environment
Encountered an issue where node-sass has some configurations specific
to the environment the project exists in. There is a conflict with
the Mac OS mounting its directory in the linux server. I am going to
go ahead and commit this file in progress as I need to focus on the
frontend for now. I will be able to fix this problem later using
docker compose.